# doc/package-manager/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "package-manager API",
    "version" : "0.2.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/api/functions" : {
      "get" : {
        "tags" : [ "Function Resource" ],
        "parameters" : [ {
          "name" : "desc",
          "in" : "query",
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationOFunction_Public"
                }
              }
            }
          }
        }
      }
    },
    "/api/functions/{funcKey}" : {
      "get" : {
        "tags" : [ "Function Resource" ],
        "parameters" : [ {
          "name" : "funcKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OFunction_Public"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Function Resource" ],
        "parameters" : [ {
          "name" : "funcKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OFunction_Public"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DSMap_Public" : {
        "type" : "array",
        "items" : {
          "type" : "string"
        }
      },
...
      "DataflowStep_Public" : {
        "type" : "object",
        "properties" : {
          "function" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "as" : {
            "type" : "string"
          },
          "inputRefs" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "args" : {
            "$ref" : "#/components/schemas/DSMap_Public"
          },
          "argRefs" : {
            "$ref" : "#/components/schemas/DSMap_Public"
          }
        }
      },
...
      "DeploymentCondition_Public" : {
        "enum" : [ "PENDING", "DEPLOYING", "RUNNING", "DOWN", "DELETED" ],
        "type" : "string"
      },
...
      "FunctionState_Public" : {
        "enum" : [ "ENABLED", "DISABLED", "REMOVING" ],
        "type" : "string"
      },
...
      "FunctionType_Public" : {
        "enum" : [ "TASK", "IM_TASK", "LOGICAL", "MACRO", "STATIC", "READONLY", "STATIC_READONLY" ],
        "type" : "string"
      },
...
      "KDeploymentProvision_Public" : {
        "type" : "object",
        "properties" : {
          "image" : {
            "type" : "string"
          },
          "replicas" : {
            "format" : "int32",
            "type" : "integer"
          },
          "requestsCpu" : {
            "type" : "string"
          },
          "requestsMemory" : {
            "type" : "string"
          },
          "limitsCpu" : {
            "type" : "string"
          },
          "limitsMemory" : {
            "type" : "string"
          },
          "env" : {
            "$ref" : "#/components/schemas/DSMap_Public"
          },
          "apiPath" : {
            "type" : "string"
          },
          "port" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "KnativeProvision_Public" : {
        "type" : "object",
        "properties" : {
          "image" : {
            "type" : "string"
          },
          "minScale" : {
            "format" : "int32",
            "type" : "integer"
          },
          "maxScale" : {
            "format" : "int32",
            "type" : "integer"
          },
          "concurrency" : {
            "format" : "int32",
            "type" : "integer"
          },
          "targetConcurrency" : {
            "format" : "int32",
            "type" : "integer"
          },
          "scaleDownDelay" : {
            "type" : "string"
          },
          "requestsCpu" : {
            "type" : "string"
          },
          "requestsMemory" : {
            "type" : "string"
          },
          "limitsCpu" : {
            "type" : "string"
          },
          "limitsMemory" : {
            "type" : "string"
          },
          "env" : {
            "$ref" : "#/components/schemas/DSMap_Public"
          },
          "apiPath" : {
            "type" : "string"
          },
          "port" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "MacroSpec_Public" : {
        "type" : "object",
        "properties" : {
          "steps" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DataflowStep_Public"
            }
          },
          "exports" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WorkflowExport_Public"
            }
          },
          "export" : {
            "type" : "string"
          },
          "atomic" : {
            "type" : "boolean"
          }
        }
      },
...
      "OFunctionDeploymentStatus_Public" : {
        "type" : "object",
        "properties" : {
          "condition" : {
            "$ref" : "#/components/schemas/DeploymentCondition_Public"
          },
          "invocationUrl" : {
            "type" : "string"
          },
          "errorMsg" : {
            "type" : "string"
          }
        }
      },
      "OFunction_Public" : {
        "required" : [ "name", "type", "outputCls" ],
        "type" : "object",
        "properties" : {
          "_key" : {
            "type" : "string"
          },
          "name" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "pkg" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/FunctionType_Public"
          },
          "outputCls" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "macro" : {
            "$ref" : "#/components/schemas/MacroSpec_Public"
          },
          "provision" : {
            "$ref" : "#/components/schemas/ProvisionConfig_Public"
          },
          "variableDescriptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VariableDescription_Public"
            }
          },
          "status" : {
            "$ref" : "#/components/schemas/OFunctionDeploymentStatus_Public"
          },
          "state" : {
            "$ref" : "#/components/schemas/FunctionState_Public"
          }
        }
      },
...
      "PaginationOFunction_Public" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "format" : "int64",
            "type" : "integer"
          },
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "itemCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OFunction_Public"
            }
          }
        }
      },
...
      "ProvisionConfig_Public" : {
        "type" : "object",
        "properties" : {
          "knative" : {
            "$ref" : "#/components/schemas/KnativeProvision_Public"
          },
          "deployment" : {
            "$ref" : "#/components/schemas/KDeploymentProvision_Public"
          },
          "staticUrl" : {
            "$ref" : "#/components/schemas/StaticUrlProvision_Public"
          }
        }
      },
...
      "StaticUrlProvision_Public" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          }
        }
      },
...
      "VariableDescription_Public" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "comment" : {
            "type" : "string"
          }
        }
      },
...
      "WorkflowExport_Public" : {
        "type" : "object",
        "properties" : {
          "from" : {
            "type" : "string"
          },
          "as" : {
            "type" : "string"
          }
        }
      }
    }
  }
}